ref(pkg/catalog): split routes file
+ splits routes file into file for inbound policies
and file for outbound policies.
+ delete routes.go and routes_test.go files
+ makes test files smaller and easier to read
+ mv wildCardRouteMatch var to pkg/trafficpolicy
